Delay cleanup of the ComputeDomain config dir on daemon cleanup
We now only remove this directory once the ComputeDomain has been
deleted (rather than on each "unprepare" of the device representing the
compute domain daemon).

There is a (rare) chance when a pod gets force deleted and a new pod
associated with the same compute domain gets started that deleting this
here will occur *after* the creation of the new pod, rendering this
directory invalid when trying to be used by the new pod. For now, we
just wait for the cleanup loop to take care of this, but in the future
we should do this cleanup on ComputeDomain deletion (in addition to the
cleanup loop).

Signed-off-by: Kevin Klues <kklues@nvidia.com>