gpu plugin: PU lock: change timeout from 10 to 300 seconds
Under alloc/dealloc pressure in the context of dynamic MIG
device allocation, it is apparent that requests line up behind
this lock.

With four physical GPUs and ~7 MIG devices per GPU, there are 28
devices to be managed. If each of these devices runs a job that is
expected to have a duration of ~1 minute, then there are

~30 Prepare()s per minute
~30 Unpepare()s per minute

That leads up to one required Prep/Unprep operation per second.

Now, it becomes apparent that each of these operations may last
longer than a second.

In any case, I have under pressure seen the PU lock acquisition
to frequently time out, in which case the same action is going
to be retried _later_, potentially much later. The system converges
faster if we just leave these requests _lined up_ (in order) and
process them as quickly as we can. Hence, I believe it certainly
makes sense to bump this timeout constant to northwards the retrying
constant at which the kubelet would retry the Prepare() request anyway.

Signed-off-by: Dr. Jan-Philip Gehrcke <jgehrcke@nvidia.com>