build: use curl w/ CI-oriented flags, zstd-compress layers
TODO: use gzip for tags? (proper releases)

The general wisdom is that both compression
and decompression for individual layers
(not just the final image!) can consume
much less time with zstd over classical
gzip. This is extremely valuable in CI,
especially for
- the flow where before a test can start
  an image needs to be pushed and pulled
- a GHCR image layer cache that allows
  for cross-build/cross-agent layer
  caching (this is about data _not_
  ending up in the prod image).

Here I also remove the variable ATTACH_ATTESTATIONS
because we do not need that degree of freedom for now.
It was recently introduced but is _not_ used.

Signed-off-by: Dr. Jan-Philip Gehrcke <jgehrcke@nvidia.com>